fix: edit data type

+++ database/migrations/2024_03_19_010736_review.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('review', function (Blueprint $table){
         
             $table->id();
-            $table->integer('user_id');
-            $table->integer('book_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('book_id');
             $table->string('title');
             $table->text('content');
